[BE] refactor(#132): 예외 추가 및 위치 변경
