+++ backend/src/main/java/com/example/backend/common/exception/study/StudyInfoException.java
@@ -1,0 +1,11 @@
+package com.example.backend.common.exception.study;
+
+import com.example.backend.common.exception.ExceptionMessage;
+import com.example.backend.common.exception.GitudyException;
+
+
+public class StudyInfoException extends GitudyException {
+    public StudyInfoException(ExceptionMessage message) {
+        super(message.getText());
+    }
+}
