+++ backend/src/test/java/com/example/backend/domain/define/study/StudyCategory/info/StudyCategoryFixture.java
@@ -1,4 +1,4 @@
-package com.example.backend.domain.define.study.StudyCategory;
+package com.example.backend.domain.define.study.StudyCategory.info;
 
 import com.example.backend.domain.define.study.category.info.StudyCategory;
